+++ Datascience/models/ModelConfigurationDetails.cs
@@ -47,5 +47,11 @@ namespace Oci.DatascienceService.Models
         [JsonProperty(PropertyName = "bandwidthMbps")]
         public System.Nullable<int> BandwidthMbps { get; set; }
         
+        /// <value>
+        /// The maximum network bandwidth for the model deployment.
+        /// </value>
+        [JsonProperty(PropertyName = "maximumBandwidthMbps")]
+        public System.Nullable<int> MaximumBandwidthMbps { get; set; }
+        
     }
 }